+++ bastion/__init__.py
@@ -1,4 +1,4 @@
 """
 Bastion Firewall Package
 """
-__version__ = '2.0.27'
+__version__ = '2.0.28'